Another compile error on non-GLIBC :(
Was seen to break on FreeBSD, Alpine and MacOSX

+++ src/mc/api/states/State.cpp
@@ -260,8 +260,8 @@ void intrusive_ptr_add_ref(State* state)
 
 void intrusive_ptr_release(State* state)
 {
-  XBT_DEBUG("[tid : %s] Removing a ref to state #%ld, %ld ref remaining", xbt::gettid().c_str(), state->get_num(),
-            state->refcount_.load());
+  XBT_DEBUG("[tid : %s] Removing a ref to state #%ld, %d ref remaining", xbt::gettid().c_str(), state->get_num(),
+            static_cast<int>(state->refcount_.load()));
   if (state->refcount_.fetch_sub(1, std::memory_order_release) == 1) {
     std::atomic_thread_fence(std::memory_order_acquire);
     delete state;
